[#5] Moving AkvoResponseGrouper tests
