+++ dev/backend/requirements.txt
@@ -1,5 +1,4 @@
-alembic==1.9.1
-anyio==3.6.2
+alembic==1.6.5
 asgi-lifespan==2.0.0
 attrs==22.2.0
 certifi==2022.12.7
@@ -19,11 +18,10 @@ iniconfig==2.0.0
 Mako==1.2.4
 MarkupSafe==2.1.1
 packaging==23.0
-pluggy==1.0.0
 psycopg2-binary==2.9.5
 pydantic==1.10.4
-pytest==7.2.0
-pytest-asyncio==0.20.3
+pytest==6.2.4
+pytest-asyncio==0.15.1
 python-dateutil==2.8.2
 python-dotenv==0.21.0
 PyYAML==6.0
